requirements.txt
# Core ML and Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Frontend
streamlit==1.28.2
plotly==5.17.0
altair==5.2.0

# Data Collection
requests==2.31.0
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# Development and Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Deployment
gunicorn==21.2.0